requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.5
Django==3.2
django-storages==1.14.2
djangorestframework==3.14.0
paramiko==3.3.1
Pillow==10.0.1
psycopg2==2.9.7
pycparser==2.21
PyNaCl==1.5.0
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0
